Add Bazel BUILD.
There is no WORKSPACE file because this BUILD file depends on
//third_party/cc/pcre:pcre.

This package has two public targets:
    swig: The SWIG compiler.
    swigwarn_swg: The rule generating the Lib/swigwarn.swg file.

Ideally, only 'swig' should be public and 'swig' should depend on 'swigwarn_swg'
in the 'data' attribute.  However, for some reason I have yet to find out,
'swigwarn_swg', when specified as part of the 'data' keyword argument for
'swig', is not treated as a dependency, i.e., does not show up in any of the
.runfiles.  Given that, in order to be able to invoke 'swig' in the
py_swig_library() rule, we have to pass in 'swigwarn_swg' as an additional
dependency to the 'swig_library' rule.  Therefore, 'swigwarn_swg' must be made
public.

TODO:
    Find out why Bazel does not honor labels specified in the 'data' keyword
    argument.  Once that is fixed, we should be able to move on.